Rename files accordingly
