+++ src/ReactComponentTemplate.js
