+++ src/index.js
@@ -1,6 +1,6 @@
 'use strict';
 
 // Babel6 does not hack the default behaviour of ES Modules anymore, so we should export
-const ReactComponentTemplate = require('./Component').default;
+const ReactComponentTemplate = require('./ReactComponentTemplate').default;
 
 module.exports = ReactComponentTemplate;
