+++ test-e2e/ReactComponentTemplate.js
